backend/requirements.txt
agno==2.3.11
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
curl-cffi==0.13.0
ddgs==9.9.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
fake-useragent==2.2.0
fastapi==0.124.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
frozendict==2.4.7
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jinja2==3.1.6
jiter==0.12.0
lxml==6.0.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
multitasking==0.0.12
numpy==2.3.5
openai==2.11.0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.1
primp==0.15.0
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
sqlalchemy==2.0.45
starlette==0.50.0
tqdm==4.67.1
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
yfinance==0.2.66
